+++ src/list.h
@@ -1,6 +1,8 @@
 #ifndef COOKIE_LIST_H
 #define COOKIE_LIST_H
 
+#include <stddef.h>
+
 /*
  * list class
  * similar to the STL list except with less features
@@ -33,7 +35,7 @@ private:
     ListNode(int v);
   } sentinel;
 
-  unsigned list_size;
+  size_t list_size;
 
 public:
   class iterator
@@ -55,17 +57,24 @@ public:
   };
 
 public:
+  /* iterator to first element of list */
   iterator begin();
+  /* iterator to end of list (note: not last element!) */
   iterator end();
 
-  unsigned size();
+  /* number of elements in this list */
+  size_t size();
+  /* is the list empty? */
   bool empty();
 
+  /* first element in this list */
   int front();
+  /* last element in this list */
   int back();
   /* get the i-th element in this list */
   int at(int i);
 
+  /* clear the contents of this list */
   void clear();
 
   /* insert a value `v` before `pos`. return a new iterator pointing to `v` */
@@ -74,9 +83,13 @@ public:
   /* erase the element `it` points to. return the iterator after the removed element */
   iterator erase(iterator it);
 
+  /* add a value to the front of the list */
   void push_front(int v);
+  /* add a value to the back of the list */
   void push_back(int v);
+  /* remove the first element of the list */
   void pop_front();
+  /* remove the last element of the list */
   void pop_back();
 };
